test-requirements.txt
django_nose==1.4.4
